requirements.txt
# --- Core API ---
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
PyJWT>=2.8
argon2-cffi>=23.1
python-dateutil>=2.9
structlog>=24.1

# --- Database & Migrations (MySQL async) ---
SQLAlchemy>=2.0
asyncmy>=0.2
alembic>=1.13

# --- Caching / Storage (optional but useful) ---
redis>=5.0
boto3>=1.34

# --- Parsing / NLP / Forecasting (Phase 1 offline) ---
regex>=2024.4.16
numpy>=1.26,<3
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4
joblib>=1.3
statsmodels>=0.14
